+++ src/main/java/lab1/tbd/serviciovoluntariado/repositories/InstitucionRepository.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface InstitucionRepository {
     public int countInstitucion();
     public List<Institucion> getAllInstitucion();
-    public Institucion getInstitucionById(int id);
+    public Institucion getInstitucionById(Long id);
     public Institucion createInstitucion(Institucion i);
     public Institucion updateInstitucion(Institucion i);
-    public void deleteInstitucionById(int id);
+    public void deleteInstitucionById(Long id);
 }
