+++ src/main/java/lab1/tbd/serviciovoluntariado/repositories/RankingRepository.java
@@ -7,9 +7,9 @@ import java.util.List;
 public interface RankingRepository {
     public int countRanking();
     public List<Ranking> getAllRanking();
-    public Ranking getRankingById(int id);
-    public List<Ranking> getRankingByTareaId(int id);
+    public Ranking getRankingById(Long id);
+    public List<Ranking> getRankingByTareaId(Long id);
     public Ranking createRanking(Ranking r);
     public Ranking updateRanking(Ranking r);
-    public void deleteRankingById(int id);
+    public void deleteRankingById(Long id);
 }
